refactor(request_service): extract PreloadRunner from PuppetCommander
Move wait/click/scroll preload choreography and interaction-budget
consumption into PreloadRunner. Commander #call delegates preload after
navigation and keeps page lifecycle ownership.
