+++ lib/html2rss/request_service/puppet_commander.rb
@@ -4,7 +4,7 @@ module Html2rss
   class RequestService
     ##
     # Commands the Puppeteer Browser to the website and builds the Response.
-    class PuppetCommander # rubocop:disable Metrics/ClassLength
+    class PuppetCommander
       BROWSER_UNSAFE_HEADERS = %w[
         host connection content-length transfer-encoding
         sec-fetch-dest sec-fetch-mode sec-fetch-site sec-fetch-user
@@ -21,9 +21,9 @@ module Html2rss
                      referer: [ctx.url.scheme, ctx.url.host].join('://'))
         @ctx = ctx
         @browser = browser
-        @skip_request_resources = skip_request_resources
         @referer = referer
         @navigation_guards = NavigationGuards.new(ctx:, skip_request_resources:)
+        @preload_runner = PreloadRunner.new(ctx:)
       end
 
       ##
@@ -33,7 +33,7 @@ module Html2rss
       def call
         page = new_page
         navigation_response = navigate_to_destination(page, ctx.url)
-        perform_preload(page)
+        preload_runner.call(page)
         navigation_guards.raise_deferred_error!
         final_navigation_response = navigation_guards.latest_navigation_response || navigation_response
         navigation_guards.validate_final!(final_navigation_response)
@@ -83,7 +83,7 @@ module Html2rss
 
       private
 
-      attr_reader :ctx, :browser, :skip_request_resources, :referer, :navigation_guards
+      attr_reader :ctx, :browser, :referer, :navigation_guards, :preload_runner
 
       def navigation_timeout_ms
         timeout = ctx.budget.remaining_timeout_seconds || ctx.policy.total_timeout_seconds
@@ -108,65 +108,6 @@ module Html2rss
       def response_url(navigation_response, fallback_url)
         raw_url = navigation_response&.url || fallback_url.to_s
         Html2rss::Url.from_absolute(raw_url)
-      end
-
-      def perform_preload(page)
-        preload_config = ctx.browserless_preload
-        return unless preload_config
-
-        wait_after(page, preload_config[:wait_after_ms])
-        click_selectors(page, preload_config[:click_selectors]) if preload_config[:click_selectors]
-        scroll_down(page, preload_config[:scroll_down]) if preload_config[:scroll_down]
-        wait_after(page, preload_config[:wait_after_ms])
-      end
-
-      def wait_after(page, timeout_ms)
-        return unless timeout_ms
-
-        ctx.budget.consume_interaction!
-        page.wait_for_timeout(timeout_ms)
-      end
-
-      def click_selectors(page, selectors)
-        selectors.each { |selector_config| click_selector(page, selector_config) }
-      end
-
-      def scroll_down(page, config)
-        iterations = config.fetch(:iterations, 1)
-        wait_after_ms = config[:wait_after_ms]
-        previous_height = nil
-
-        iterations.times do
-          updated_height = perform_scroll_iteration(page, wait_after_ms, previous_height)
-          break unless updated_height
-
-          previous_height = updated_height
-        end
-      end
-
-      def click_selector(page, config)
-        selector = config.fetch(:selector)
-        max_clicks = config.fetch(:max_clicks, 1)
-        wait_after_ms = config[:wait_after_ms]
-
-        max_clicks.times do
-          break unless (element = page.query_selector(selector))
-
-          ctx.budget.consume_interaction!
-          element.click
-          wait_after(page, wait_after_ms)
-        end
-      end
-
-      def perform_scroll_iteration(page, wait_after_ms, previous_height)
-        ctx.budget.consume_interaction!
-        page.evaluate('() => window.scrollTo(0, document.body.scrollHeight)')
-        wait_after(page, wait_after_ms)
-
-        current_height = page.evaluate('() => document.body.scrollHeight')
-        return if previous_height && current_height <= previous_height
-
-        current_height
       end
     end
   end
